refactor: add containers and change styles
