+++ src/components/CourseInfo/components/CourseInfoCard/styles.module.scss
@@ -2,7 +2,7 @@
     display: grid;
     grid-template-columns: auto auto auto;
     align-items: start;
-    justify-content: center;
+    justify-content: space-between;
     gap: 32px;
     padding: 53px 64px;
     border: 1px solid gray;
