+++ src/components/Courses/styles.module.scss
@@ -11,7 +11,13 @@
     justify-content: center;
     align-items: center;
     flex-direction: column;
-    
+
+    &__wrapper {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        height: 100%;
+    }
     &__content {
         display: flex;
         justify-content: center;
