+++ src/components/Header/Header.tsx
@@ -1,23 +1,10 @@
 import { Link } from 'react-router-dom';
 
-import { userLogoutThunk } from 'src/store';
-import { useAppDispatch } from 'src/hooks';
-import { ReduxContainer } from 'src/containers';
-import { Button } from 'src/common';
-import { HeaderWrapper, Logo, Username } from './components';
+import { HeaderWrapper, Logo } from './components';
+import { LogoutButtonContainer, UsernameContainer } from './containers';
 import styles from './styles.module.scss';
 
 export const Header = () => {
-	const dispatch = useAppDispatch();
-
-	const handleLogout = async () => {
-		try {
-			await dispatch(userLogoutThunk());
-		} catch (err) {
-			console.log(err);
-		}
-	};
-
 	return (
 		<header className={styles.header}>
 			<HeaderWrapper
@@ -28,22 +15,8 @@ export const Header = () => {
 				}
 				right={
 					<div className={styles.header__right}>
-						<ReduxContainer
-							selector={(state) => state.user}
-							render={({ data }) =>
-								data.isAuth && (
-									<>
-										{!!data.name && <Username name={data.name} />}
-										<Button
-											style={{ textTransform: 'uppercase' }}
-											onClick={handleLogout}
-										>
-											Logout
-										</Button>
-									</>
-								)
-							}
-						/>
+						<UsernameContainer />
+						<LogoutButtonContainer />
 					</div>
 				}
 			/>
