+++ src/components/Header/components/LogoutButton/index.ts
@@ -1,0 +1,1 @@
+export * from './LogoutButton';
