+++ src/components/Header/components/index.ts
@@ -1,3 +1,4 @@
 export * from './HeaderWrapper';
 export * from './Logo';
 export * from './Username';
+export * from './LogoutButton';
