+++ src/components/Header/containers/index.ts
@@ -1,0 +1,2 @@
+export * from './LogoutButtonContainer';
+export * from './UsernameContainer';
